Allow 'ls' and 'rm' for file, even if it doesn't exist in file system
It is sometimes necessary to list or remove files that have an entry in
the index (fixity database), but which do not exist (any longer) in the
file system. Prior to this commit, a dummy file must be created at the
desired path to enable these operations. This is not even possible when
that part of the file system is not writeable. And when it is, this
creates a risk of leaving the dummy file behind or clobbering another file.